adding types to all base plugins, moving video qualifier to abr plugin, composing plugin types
