+++ packages/url-loader/package.json
@@ -33,9 +33,9 @@
     "jest": "^29.4.0",
     "semantic-release": "^20.1.0",
     "semantic-release-monorepo": "^7.0.5",
-    "ts-jest": "^29.0.5",
-    "tsup": "^5.10.1",
-    "typescript": "^4.5.3"
+    "ts-jest": "^29.1.1",
+    "tsup": "^8.0.1",
+    "typescript": "^5.3.3"
   },
   "publishConfig": {
     "access": "public"
@@ -44,7 +44,7 @@
     "@cloudinary-util/util": "workspace:*",
     "@cloudinary/url-gen": "^1.10.2",
     "@semantic-release/npm": "^9.0.2",
-    "zod": "^3.21.4",
-    "zod-to-json-schema": "^3.21.1"
+    "zod": "^3.22.4",
+    "zod-to-json-schema": "^3.22.3"
   }
 }
