+++ packages/url-loader/src/index.ts
@@ -5,7 +5,7 @@ export { effects, position, primary, text } from './constants/qualifiers';
 
 export type { AssetOptions } from './types/asset';
 export type { ImageOptions } from './types/image';
-export type { VideoOptionsResize, VideoOptions } from './types/video';
+export type { VideoOptions } from './types/video';
 
 export type { AnalyticsOptions } from './types/analytics';
 export type { ConfigOptions } from './types/config';
