+++ packages/url-loader/src/plugins/remove-background.ts
@@ -1,7 +1,18 @@
+import { z } from 'zod';
+
 import { ImageOptions } from '../types/image';
 import { PluginSettings } from '../types/plugins';
 
-export const props = ['removeBackground'];
+export const pluginProps = {
+  removeBackground: z.boolean()
+    .describe(JSON.stringify({
+      text: 'Removes the background of an image using the Cloudinary AI Background Removal Add-On (Required).',
+      url: 'https://cloudinary.com/documentation/cloudinary_ai_background_removal_addon'
+    }))
+    .optional(),
+};
+
+export const props = Object.entries(pluginProps).map(([name]) => name);
 export const assetTypes = ['image', 'images'];
 
 export function plugin(props: PluginSettings<ImageOptions>) {
