+++ packages/url-loader/src/plugins/seo.ts
@@ -1,6 +1,17 @@
+import { z } from 'zod';
+
 import { PluginSettings } from '../types/plugins';
 
-export const props = ['seoSuffix'];
+export const pluginProps = {
+  seoSuffix: z.string()
+    .describe(JSON.stringify({
+      text: 'Configures the URL to include an SEO-friendly suffix in the URL',
+      url: 'https://cloudinary.com/documentation/advanced_url_delivery_options#seo_friendly_media_asset_urls'
+    }))
+    .optional()
+};
+
+export const props = Object.entries(pluginProps).map(([name]) => name);
 export const assetTypes = ['image', 'images', 'video', 'videos'];
 
 export function plugin(props: PluginSettings) {
