+++ packages/url-loader/src/plugins/version.ts
@@ -1,6 +1,20 @@
+import { z } from 'zod';
+
 import { PluginSettings } from '../types/plugins';
 
-export const props = ['version'];
+export const pluginProps = {
+  version: z.union([
+      z.number(),
+      z.string()
+    ])
+    .describe(JSON.stringify({
+      text: 'Custom version number to apply to asset URL.',
+      url: 'https://cloudinary.com/documentation/advanced_url_delivery_options#asset_versions'
+    }))
+    .optional()
+};
+
+export const props = Object.entries(pluginProps).map(([name]) => name);
 export const assetTypes = ['image', 'images', 'video', 'videos'];
 
 export function plugin(props: PluginSettings) {
