+++ packages/url-loader/src/types/video.ts
@@ -2,21 +2,12 @@ import { z } from 'zod';
 
 import { assetOptionsSchema } from './asset';
 
-// Resize
+import { pluginProps as abrPluginProps } from '../plugins/abr';
 
-export const videoOptionsResizeSchema = assetOptionsSchema;
-
-export type VideoOptionsResize = z.infer<typeof videoOptionsResizeSchema>;
-
-// Video Options
-
-export const videoOptionsSchema = assetOptionsSchema.extend({
-  streamingProfile: z.string()
-    .describe(JSON.stringify({
-      text: 'The streaming profile to apply when delivering a video using adaptive bitrate streaming.',
-      url: 'https://cloudinary.com/documentation/transformation_reference#sp_streaming_profile'
-    }))
-    .optional(),
-})
+export const videoOptionsSchema = assetOptionsSchema.merge(z.object({
+  // Spreading plugins instead of extend or merge to avoid excessive schema warning
+  // https://github.com/microsoft/TypeScript/issues/34933#issuecomment-1772787785
+  ...abrPluginProps,
+}))
 
 export type VideoOptions = z.infer<typeof videoOptionsSchema>;
